Run TypeMap logic on demand (#1728)
* Run TypeMap logic on demand

* PR feedback

- Move default interface implementations to TypeMapInfo
- Avoid small helpers
